Rename folder to match the gem name.
