+++ ffmpeg-screenrecorder.gemspec
@@ -1,6 +1,6 @@
 lib = File.expand_path('lib', __dir__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'ffmpeg/version'
+require 'ffmpeg-screenrecorder/version'
 
 Gem::Specification.new do |spec|
   spec.name        = 'ffmpeg-screenrecorder'
