+++ lib/ffmpeg-screenrecorder/recorder_options.rb
