+++ lib/ffmpeg-screenrecorder/version.rb
@@ -1,0 +1,5 @@
+module FFMPEG
+  class ScreenRecorder
+    VERSION = '1.0.0.beta10'.freeze
+  end
+end
