+++ lib/ffmpeg-screenrecorder/window_titles.rb
