--- lib/ffmpeg/version.rb
@@ -1,5 +1,0 @@
-module FFMPEG
-  class ScreenRecorder
-    VERSION = '1.0.0.beta10'.freeze
-  end
-end
